Hoist most of multiprocess tests up to common base class
